Adapt JsonLines to use the new streaming code.
JsonLines::fromStreamOnce() was a casualty, but there are no known uses of
it in the wild (it was used internally by the old fromStream), and
it would be weird for it to be kicking around with slightly different
semantics and less resiliency. (The new code does not, strictly speaking,
require all the JSON to be on one line. It only requires that there is
at least a newline or carriage return between each element.